pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.django-stubs]
django_settings_module = "travel_webapp.settings"
strict_settings = false

[tool.mypy]
ignore_missing_imports = true
exclude = '''
/(
    | migrations
)/
'''
plugins = ["mypy_django_plugin.main"]


[tool.poetry]
name = "travel-webapp"
version = "0.1.0"
description = ""
authors = ["Zack Kuptsow"]
license = "GNU AFFERO GENERAL PUBLIC LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
django = "^5.0.3"
mypy = "^1.9.0"
pytest = "^7.1.3"
python = "^3.11.5"
django-simple-history = "^3.5.0"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
django-constance = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.8.0"
django-stubs = "^4.2.7"

[tool.pydocstyle]
convention = "pep257"
add-ignore = ["D100", "D104", "D106", "D203", "D212", "D406", "D407", "D411", "D413"]
add-select = ["D400"]
match = '(.*\/migrations\/.*) | ((?!apps).)*\.py'

[tool.pytest.ini_options]
minversion = "7.1.3"
DJANGO_SETTINGS_MODULE = "travel_webapp.settings"
testpaths = [
    "activity/tests",
    "base/tests",
]
python_files = [
    "tests.py",
    "test_*.py",
    "*_tests.py"
]
django_find_project = true
addopts = "--strict-markers"


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]


# Same as Black.
line-length = 88
indent-width = 4

target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"
